Merge Rezyn's buildlog and revert from http://rezyn.mercenariesguild.net/aardvark/aardvark_0303_buildlog.patch for testing
The Domination patch generalises the build point queue code; it no longer assumes that repeaters are the only buildable that give BP zones and that humans are the only team that can have BP zones.  The code was patched by Rezyn's buildlog-and-revert patch and was modified to be compatible with the BP zone generalisation.