client: reset class overrides when OC mode changes.
Thanks for reporting this issue.  This commit should fix this particular one.

It seemed BG_InitBuildableConfigs() etc. were called when the config string
was not set with OC mode enabled yet, and then the OC mode was set by the
client afterwards, but the initialization was already done and wasn't being
reset.